Removed Room

+++ app/src/main/java/com/gmail/bogumilmecel2/fitnessappv2/feature_diary/domain/use_cases/CalculateSkipUseCase.kt
@@ -1,5 +1,5 @@
 package com.gmail.bogumilmecel2.fitnessappv2.feature_diary.domain.use_cases
 
 class CalculateSkipUseCase {
-    operator fun invoke(page: Int, sizePerPage: Int) = sizePerPage * (page - 1)
+    operator fun invoke(page: Int, sizePerPage: Long) = sizePerPage * (page - 1)
 }
