+++ app/src/main/java/com/gmail/bogumilmecel2/fitnessappv2/feature_diary/domain/use_cases/ShouldDisplayNextPageUseCase.kt
@@ -1,7 +1,7 @@
 package com.gmail.bogumilmecel2.fitnessappv2.feature_diary.domain.use_cases
 
 class ShouldDisplayNextPageUseCase {
-    operator fun invoke(size: Int, perPage: Int, currentPage: Int): Boolean {
+    operator fun invoke(size: Int, perPage: Long, currentPage: Int): Boolean {
         return size >= currentPage * perPage
     }
 }
